+++ src/server/auth/index.js
@@ -1,7 +1,7 @@
 import { loginController } from '~/src/server/auth/login.js'
 import { signInOidcController } from '~/src/server/auth/sign-in-oidc.js'
 import { signOutController } from '~/src/server/auth/sign-out.js'
-import { loginEntraController } from '~/src/server/auth/login-entra.js'
+import { loginEntraController } from '~/src/server/auth/sign-in-entra.js'
 import { signInOidcEntraController } from '~/src/server/auth/sign-in-oidc-entra.js'
 
 export const auth = {
