+++ src/server/auth/login.test.js
@@ -17,7 +17,7 @@ describe('#loginController', () => {
   test('should render the project name page', async () => {
     const { statusCode, headers } = await server.inject({
       method: 'GET',
-      url: routes.LOGIN
+      url: routes.SIGNIN
     })
 
     expect(statusCode).toBe(statusCodes.redirect)
