+++ src/server/common/constants/routes.js
@@ -15,8 +15,8 @@ export const routes = {
   ACTIVITY_DESCRIPTION: '/exemption/activity-description',
   AUTH_DEFRA_ID_CALLBACK: '/signin-oidc',
   AUTH_ENTRA_ID_CALLBACK: '/auth',
-  LOGIN: '/login',
-  LOGIN_ENTRA: '/login-entra',
+  SIGNIN: '/signin',
+  SIGNIN_ENTRA: '/signin-entra',
   SIGN_OUT: '/sign-out',
   WIDTH_OF_SITE: '/exemption/width-of-site',
   CHOOSE_FILE_UPLOAD_TYPE: '/exemption/choose-file-type-to-upload',
